requirements.txt
pytest
pylint
black
tensorflow
tensorflow-hub
tensorflow-text
tensorflow-ranking
tensorflow-recommenders
tensorflow-addons
pandas
tqdm
scikit-learn
apache-airflow # Needs Apache Airflow installed
apache-airflow-providers-airbyte # Needs Airbyte installed
bs4

requests
pyhtml2pdf
txt2pdf.core
pdfy
pikepdf
pypdf
glob
os
shutil
concurrent.futures
threading
argparse
datetime
psutil